examples: dual_qemu_ivshmem: add RPMsg over IVSHMEM sample code.
The additions include a backend to glue the Zephyr IVSHMEM
device driver into the openAMP code making it usable to send
data between two QEMU instances using the RPMsg protocol.

Also a custom shell command in the host side application
is provided to send string messages for a number of times.

Signed-off-by: Felipe Neves <felipe.neves@linaro.org>