examples: Add support for baremetal R5 in the TI K3 family devices
Adding baremetal support for the R5 to use RPMsg for the devices
in the TI K3 family, specifically for the SK-AM64. Users should
be able to replace memory addresses for the R5 in the
linker_remote.ld file and platform_info.h files to make RPMsg
compatible with the R5 in their chosen board. Cmake file was
added to aid in building for the R5 in TI K3 family devices.
Mailbox driver added to allow RPMsg to work with the TI K3 family
devices. Other API files added to aid in interrupts and other
board specific functions. IPI and no IPI is supported in the
platform_info files to allow mailbox to kick RPMsg. Matrix
multiply example is functional as a demo.

Signed-off-by: Kendall Willis <kendallwillis@ti.com>