west.yml: Clone our libraries and their zephyr glue
This manifest allows us to use our repos for the open-amp and libmetal
libraries.  As our libraries do not have the zephyr dir, they are not
seem as zephyr modules.

To make this work easily with zephyr we also add a zephyr glue layer.
This layer _is_ seen as a zephyr module and integrates both libraries.

Several approaches were tried for this and this was the chosen method for
the following reasons:

Pros:
* All needed files are provided by just doing the west update
* We don't need to set any ENV vars each time like EXTRA_ZEPHYR_MODULES
* It works with our samples and zephyr stock samples
* The version of open-amp and libmetal are directly controlled in
  the manifest
* It is low maintenance as the module boiler plate is unlikely to change
  much.

Cons:
* One extra repo

Signed-off-by: Bill Mills <bill.mills@linaro.org>
Co-authored-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>