Xilinx DTS: Switch to all overlays
Build upstream DTBs and DTBOs (mostly) as is.
(Overlay support is added to zcu102).

Then build a series of overlays that can be applied to the upstream base.
These overlays can be used at boot time if desired.

Then go ahead and apply the overlays to the bases to produce full dtbs
so people that don't want to learn about overlays can just use the one
dtb that they want.

Signed-off-by: Bill Mills <bill.mills@linaro.org>