dtb: Add Xilinx baremetal demo compatible DTB for v6.5 kernel
The Xilinx baremetal apps as built by Xilinx v2023.2 use a different
mailbox configuration than the Zephyr examples do.

We wish to align these in the future but for now we need a different DTB.
We can not use the Xilinx v2023.2 built DTB as it is not compatible with
the upstream remoteproc driver (it is still using the pre-upstream driver.)

This change adds an overlay for this configuration and then also applies
it to make complete dtbs.  Boot configs can either apply the overlays at
boot time or use the complete dtbs.

Signed-off-by: Bill Mills <bill.mills@linaro.org>