legacy apps: update cmake minimum support to 3.16
We currently support an old version of CMake (3.0.2) that was released
in 2018. Next cmake revisions will support 3.5 as minimum version.

Update the minimum version to 3.16, aligned with the libmetal and the
open-amp libraies.

By updating to CMake 3.16 as the minimum version, we can also clean up
the following policy add-ons:
- CMP0048 added in CMake v3.0
- CMP0077 added in CMake v3.13

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>