legacy_apps: zynqmp_r5: move caching resource table call to get_resource_table
Previously this was call to restore_initial_rsc_table to ensure that
resource table snapshot was cached early enough to prevent a warning. This
should instead be call to get_resource_table so that the future
call to restore_initial_rsc_table will have the resource set up.

Otherwise the snapshot can still fail if the resource is not set up.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>
Acked-by: Tanmay Shah <tanmay.shah@amd.com>