examples: legacy_apps: Add default routines for suspend/resume
The application or the CPU can be suspended and resumed while waiting
for new RPMsg messages from the remote processor.
This commit introduces `system_suspend` and `system_resume` functions,
along with their implementations for different systems.

- FreeRTOS: The suspend/resume functionality involves suspending and
  resuming the application.
- Linux and Baremetal: By default, `metal_cpu_yield()` is called.
  However, these functions can be redefined in the machine layer if
  needed.

The application can be suspended and resume waiting
from new rpmsg from the remote.
Add system_suspend and system_resume functions and associated
implementation for the different system.
For the FreeRTOS the suspend/resume consist in suspending resuming
the application
For Linux and baremetal metal_cpu_yield() is called by default, but
some functions can be redefined in the machine on need

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>